improve code style
If one orientates themselves according to common Stylesheets (see Google Stylesheet),
there is no line break before the opening curly bracket! This may be a convention in
languages like C#, but it is not in Java and should therefor be fixed in this repository,
as students who are learning Java will probably adapt to the conventions used in those tests.