[W08H02] test(TaskFunctionBehaviorTest): Add strict equals comparison (#190)
> Überschriebene Methoden hashCode und equals, die jeweils nur von ID
abhängig sein sollen.

The goal of the first test is to see if TaskFunction::equals uses a
Class::getClass comparison thus locking Class::equals to instances of
the same class. TaskFunction::equals may only be dependent on ID, in
case of altering a different instance of TaskFunction to the same value
(e.g. by a method overwrite) it should yield true.

The second test is trivial as we are not overwriting any methods in the
subclass, we could still produce an undefined behaviour in flawed
implementations that can't handle subclasses.